--- PinochleDeck/PinochleDeck/PinochleDeck.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PinochleDeck", "PinochleDeck\PinochleDeck.csproj", "{E8DC97AE-4F51-495C-BD8E-0C27827C915F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E8DC97AE-4F51-495C-BD8E-0C27827C915F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E8DC97AE-4F51-495C-BD8E-0C27827C915F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E8DC97AE-4F51-495C-BD8E-0C27827C915F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E8DC97AE-4F51-495C-BD8E-0C27827C915F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
