--- PinochleDeck/PinochleDeck/PinochleDeck/PinochleBoard.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PinochleDeck
-{
-    public class PinochleBoard
-    {
-        public PinochleBoard()
-        {
-
-        }
-    }
-}
